implemented jwt authentication and authorization and update my users table schema
implemented jwt authentication to our sever for some routes and also
added admin and user authorization. For that made some changes in my
users schema, like added token,refresh token and  role column for
storing tokens and  if the user is a normal user or a admin. wrote
queries for the same and storing and updating the tokens